Move view menu logic from lgv to bookmark plugin
