+++ plugins/linear-genome-view/src/LinearGenomeView/index.tsx
@@ -39,8 +39,6 @@ import VisibilityIcon from '@material-ui/icons/Visibility'
 import LabelIcon from '@material-ui/icons/Label'
 import FolderOpenIcon from '@material-ui/icons/FolderOpen'
 import PhotoCameraIcon from '@material-ui/icons/PhotoCamera'
-import BookmarkIcon from '@material-ui/icons/Bookmark'
-import BookmarksIcon from '@material-ui/icons/Bookmarks'
 import clone from 'clone'
 import { AnyConfigurationModel } from '@jbrowse/core/configuration/configurationSchema'
 import { saveAs } from 'file-saver'
@@ -674,42 +672,6 @@ export function stateModelFactory(pluginManager: PluginManager) {
         throw new Error(`invalid track selector type ${self.trackSelectorType}`)
       },
 
-      activateBookmarkWidget() {
-        const session = getSession(self)
-        if (isSessionModelWithWidgets(session)) {
-          // @ts-ignore
-          let bookmarkWidget = session.widgets.get('GridBookmark')
-          if (!bookmarkWidget) {
-            bookmarkWidget = session.addWidget(
-              'GridBookmarkWidget',
-              'GridBookmark',
-              { view: self },
-            )
-          }
-
-          session.showWidget(bookmarkWidget)
-          return bookmarkWidget
-        }
-
-        throw new Error('Could not open bookmark widget')
-      },
-
-      bookmarkCurrentRegion() {
-        const selectedRegions = this.getSelectedRegions(
-          self.leftOffset,
-          self.rightOffset,
-        )
-        const firstRegion = selectedRegions[0]
-        // @ts-ignore
-        const { widgets } = getSession(self)
-        let bookmarkWidget = widgets.get('GridBookmark')
-        if (!bookmarkWidget) {
-          this.activateBookmarkWidget()
-          bookmarkWidget = widgets.get('GridBookmark')
-        }
-        bookmarkWidget.addBookmark(firstRegion)
-      },
-
       navToLocString(locString: string, optAssemblyName?: string) {
         const { assemblyManager } = getSession(self)
         const { isValidRefName } = assemblyManager
@@ -1271,16 +1233,6 @@ export function stateModelFactory(pluginManager: PluginManager) {
               label: 'Horizontally flip',
               icon: SyncAltIcon,
               onClick: self.horizontallyFlip,
-            },
-            {
-              label: 'Bookmark current region',
-              icon: BookmarkIcon,
-              onClick: self.bookmarkCurrentRegion,
-            },
-            {
-              label: 'Open bookmark widget',
-              icon: BookmarksIcon,
-              onClick: self.activateBookmarkWidget,
             },
             { type: 'divider' },
             {
